+++ gateway/src/main/java/org/bp/pizza/model/Utils.java
@@ -3,7 +3,7 @@ package org.bp.pizza.model;
 import java.math.BigDecimal;
 
 public class Utils {
-	static public PizzaInfo preparePizzaInfo(String pizzaId, BigDecimal cost) {
+	static public PizzaInfo preparePizzaInfo(String pizzaId, Float cost) {
 		PizzaInfo pizzaInfo = new PizzaInfo();
 		pizzaInfo.setId(pizzaId);
 		pizzaInfo.setCost(cost);
